Move App.js to /src
