+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Calculator from './Calculator';
+import Calculator from './components/Calculator';
 import '../css/App.css';
 
 const App = () => (
